Model-API/requirements.txt
fastapi
uvicorn
numpy
pillow
matplotlib==3.7.1